consolidate system classes
