+++ legate/driver/logs.py
@@ -27,10 +27,10 @@ from typing import TYPE_CHECKING, Iterator
 from ..utils.ui import warn
 
 if TYPE_CHECKING:
+    from ..utils.system import System
     from ..utils.types import Command
     from .config import Config
     from .launcher import Launcher
-    from .system import System
 
 __all__ = (
     "DebuggingHandler",
