+++ legate/tester/stages/_osx/omp.py
@@ -29,7 +29,7 @@ if TYPE_CHECKING:
     from ....utils.types import ArgList, EnvDict
     from ... import FeatureType
     from ...config import Config
-    from ...system import System
+    from ...test_system import TestSystem
 
 
 class OMP(TestStage):
@@ -40,7 +40,7 @@ class OMP(TestStage):
     config: Config
         Test runner configuration
 
-    system: System
+    system: TestSystem
         Process execution wrapper
 
     """
@@ -49,10 +49,10 @@ class OMP(TestStage):
 
     args = [CUNUMERIC_TEST_ARG]
 
-    def __init__(self, config: Config, system: System) -> None:
+    def __init__(self, config: Config, system: TestSystem) -> None:
         self._init(config, system)
 
-    def env(self, config: Config, system: System) -> EnvDict:
+    def env(self, config: Config, system: TestSystem) -> EnvDict:
         return UNPIN_ENV
 
     def shard_args(self, shard: Shard, config: Config) -> ArgList:
@@ -63,7 +63,7 @@ class OMP(TestStage):
             str(config.ompthreads),
         ]
 
-    def compute_spec(self, config: Config, system: System) -> StageSpec:
+    def compute_spec(self, config: Config, system: TestSystem) -> StageSpec:
         omps, threads = config.omps, config.ompthreads
         procs = omps * threads + config.utility
         workers = adjust_workers(
