+++ legate/tester/test_system.py
@@ -20,14 +20,15 @@ from __future__ import annotations
 
 import multiprocessing
 import os
-import sys
 from dataclasses import dataclass
-from functools import cached_property
 from pathlib import Path
 from subprocess import PIPE, STDOUT, run as stdlib_run
 from typing import Sequence
 
-from ..utils.types import CPUInfo, EnvDict, GPUInfo
+from ..utils.system import System
+from ..utils.types import EnvDict
+
+__all__ = ("TestSystem",)
 
 
 @dataclass
@@ -49,7 +50,7 @@ class ProcessResult:
     output: str = ""
 
 
-class System:
+class TestSystem(System):
     """A facade class for system-related functions.
 
     Parameters
@@ -120,51 +121,3 @@ class System:
             returncode=proc.returncode,
             output=proc.stdout,
         )
-
-    @cached_property
-    def cpus(self) -> tuple[CPUInfo, ...]:
-        """A list of CPUs on the system."""
-
-        N = multiprocessing.cpu_count()
-
-        if sys.platform == "darwin":
-            return tuple(CPUInfo((i,)) for i in range(N))
-
-        sibling_sets: set[tuple[int, ...]] = set()
-        for i in range(N):
-            line = open(
-                f"/sys/devices/system/cpu/cpu{i}/topology/thread_siblings_list"
-            ).read()
-            sibling_sets.add(
-                tuple(sorted(int(x) for x in line.strip().split(",")))
-            )
-        return tuple(CPUInfo(siblings) for siblings in sorted(sibling_sets))
-
-    @cached_property
-    def gpus(self) -> tuple[GPUInfo, ...]:
-        """A list of GPUs on the system, including total memory information."""
-
-        try:
-            # This pynvml import is protected inside this method so that in
-            # case pynvml is not installed, tests stages that don't need gpu
-            # info (e.g. cpus, eager) will proceed unaffected. Test stages
-            # that do require gpu info will fail here with an ImportError.
-            import pynvml  # type: ignore[import]
-
-            # Also a pynvml package is available on some platforms that won't
-            # have GPUs for some reason. In which case this init call will
-            # fail.
-            pynvml.nvmlInit()
-        except Exception:
-            return ()
-
-        num_gpus = pynvml.nvmlDeviceGetCount()
-
-        results = []
-        for i in range(num_gpus):
-            info = pynvml.nvmlDeviceGetMemoryInfo(
-                pynvml.nvmlDeviceGetHandleByIndex(i)
-            )
-            results.append(GPUInfo(i, info.total))
-
-        return tuple(results)
