+++ tests/unit/legate/driver/conftest.py
@@ -20,8 +20,9 @@ from typing import Any, Callable, Iterable
 
 import pytest
 
-from legate.driver import Config, Launcher, System
+from legate.driver import Config, Launcher
 from legate.driver.config import MultiNode
+from legate.utils.system import System
 
 from .util import GenConfig, GenSystem
