+++ tests/unit/legate/driver/test_driver.py
@@ -23,9 +23,9 @@ import legate.driver.driver as m
 from legate.driver.args import LAUNCHERS
 from legate.driver.command import CMD_PARTS
 from legate.driver.launcher import Launcher
-from legate.driver.system import System
 from legate.driver.util import print_verbose
 from legate.utils.colors import scrub
+from legate.utils.system import System
 from legate.utils.types import LauncherType
 
 from ...util import Capsys
