+++ tests/unit/legate/driver/test_launcher.py
@@ -20,7 +20,7 @@ import pytest
 
 import legate.driver.launcher as m
 from legate.driver.args import LAUNCHERS
-from legate.driver.system import System
+from legate.utils.system import System
 from legate.utils.types import LauncherType
 
 from ...util import powerset_nonempty
