+++ tests/unit/legate/driver/test_main.py
@@ -28,10 +28,10 @@ import legate.driver as m
 def test_main(mocker: MockerFixture) -> None:
     import legate.driver.config
     import legate.driver.driver
-    import legate.driver.system
+    import legate.utils.system
 
     config_spy = mocker.spy(legate.driver.config.Config, "__init__")
-    system_spy = mocker.spy(legate.driver.system.System, "__init__")
+    system_spy = mocker.spy(legate.utils.system.System, "__init__")
     driver_spy = mocker.spy(legate.driver.driver.Driver, "__init__")
     mocker.patch("legate.driver.driver.Driver.run", return_value=123)
 
@@ -48,7 +48,7 @@ def test_main(mocker: MockerFixture) -> None:
     assert driver_spy.call_count == 1
     assert len(driver_spy.call_args[0]) == 3
     assert isinstance(driver_spy.call_args[0][1], legate.driver.config.Config)
-    assert isinstance(driver_spy.call_args[0][2], legate.driver.system.System)
+    assert isinstance(driver_spy.call_args[0][2], legate.utils.system.System)
     assert driver_spy.call_args[1] == {}
 
     assert result == 123
