+++ tests/unit/legate/utils/sample_cmake_cache.txt
